requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
easyocr==1.7.2
eventlet==0.39.1
falcon==4.0.2
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
git-filter-repo==2.47.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mongoengine==0.29.1
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
oauthlib==3.2.2
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.2
pydantic_core==2.33.1
pymongo==4.11.3
pyparsing==3.2.3
pytesseract==0.3.13
python-bidi==0.6.6
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.25.2
scipy==1.15.2
shapely==2.1.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tifffile==2025.3.30
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
